Cheap subsampling

Method implementing the cheap subsampling method for confidence intervals

cheap_bootstrap(
  fun,
  b = 20,
  size = NULL,
  alpha = 0.05,
  keep_boot_estimates = TRUE,
  parallelize = FALSE,
  cores = parallel::detectCores(),
  data = NULL,
  progress_bar = TRUE,
  type = "subsampling"
)

Arguments

fun

A function that returns a vector of coefficients or a model object which saves the call. In the second case, a coef needs to be defined for 'class(fun)'.

b

Number of bootstrap samples.

size

Subsample size. Defaults to 0.632 * nrow(data).

alpha

Significance level. Defaults to 0.05.

keep_boot_estimates

Logical. Should the bootstrap estimates be kept? Defaults to FALSE.

parallelize

Logical. Should the bootstrap samples be computed in parallel? Defaults to FALSE.

cores

Number of cores to use for parallel computation, if parallelize = TRUE. Defaults to detectCores().

data

Data set to be used for the computation, if applicable.

progress_bar

Logical. Should a progress bar be displayed? Defaults to TRUE.

type

Character. Type of bootstrap method. Can be either "subsampling" or "non_parametric". Defaults to "subsampling".

Value

An object of class "cheap_bootstrap" containing the point estimates and confidence intervals.

Details

Given a model object or a function that returns a vector of coefficients and a data set, this function computes confidence intervals using the cheap subsampling method.

Examples

utils::data(anorexia, package = "MASS")
## example with a function call
set.seed(123)
fun <- function(data) coef(lm(Postwt ~ Prewt + Treat + offset(Prewt), data = data))
cs <- cheap_bootstrap(fun=fun, b = 20, data = anorexia)
#>   |                                                          |                                                  |   0%  |                                                          |===                                               |   5%  |                                                          |=====                                             |  11%  |                                                          |========                                          |  16%  |                                                          |===========                                       |  21%  |                                                          |=============                                     |  26%  |                                                          |================                                  |  32%  |                                                          |==================                                |  37%  |                                                          |=====================                             |  42%  |                                                          |========================                          |  47%  |                                                          |==========================                        |  53%  |                                                          |=============================                     |  58%  |                                                          |================================                  |  63%  |                                                          |==================================                |  68%  |                                                          |=====================================             |  74%  |                                                          |=======================================           |  79%  |                                                          |==========================================        |  84%  |                                                          |=============================================     |  89%  |                                                          |===============================================   |  95%  |                                                          |==================================================| 100%
cs
#> Cheap subsampling results for subsample size m = 46 and 20 bootstrap samples
#>     estimate cheap_lower cheap_upper   parameter
#> 1 49.7711090   12.115299  87.4269191 (Intercept)
#> 2 -0.5655388   -1.026930  -0.1041476       Prewt
#> 3 -4.0970655   -7.835691  -0.3584399   TreatCont
#> 4  4.5630627   -1.452799  10.5789246     TreatFT
summary(cs)
#> Cheap subsampling results for subsample size m = 46 and 20 bootstrap samples
#>     estimate cheap_lower cheap_upper   parameter
#> 1 49.7711090   12.115299  87.4269191 (Intercept)
#> 2 -0.5655388   -1.026930  -0.1041476       Prewt
#> 3 -4.0970655   -7.835691  -0.3584399   TreatCont
#> 4  4.5630627   -1.452799  10.5789246     TreatFT

## example with a model object
set.seed(123)
x <- lm(Postwt ~ Prewt + Treat + offset(Prewt), data = anorexia)
cs2 <- cheap_bootstrap(fun=x, b = 20)
#> fun is not a function, trying to retrieve call object from fun
#>   |                                                          |                                                  |   0%  |                                                          |===                                               |   5%  |                                                          |=====                                             |  11%  |                                                          |========                                          |  16%  |                                                          |===========                                       |  21%  |                                                          |=============                                     |  26%  |                                                          |================                                  |  32%  |                                                          |==================                                |  37%  |                                                          |=====================                             |  42%  |                                                          |========================                          |  47%  |                                                          |==========================                        |  53%  |                                                          |=============================                     |  58%  |                                                          |================================                  |  63%  |                                                          |==================================                |  68%  |                                                          |=====================================             |  74%  |                                                          |=======================================           |  79%  |                                                          |==========================================        |  84%  |                                                          |=============================================     |  89%  |                                                          |===============================================   |  95%  |                                                          |==================================================| 100%
summary(cs2)
#> Cheap subsampling results for subsample size m = 46 and 20 bootstrap samples
#>     estimate cheap_lower cheap_upper   parameter
#> 1 49.7711090   12.115299  87.4269191 (Intercept)
#> 2 -0.5655388   -1.026930  -0.1041476       Prewt
#> 3 -4.0970655   -7.835691  -0.3584399   TreatCont
#> 4  4.5630627   -1.452799  10.5789246     TreatFT

## example with a function call and parallel computation
if (FALSE) {
set.seed(123)
x <- function(d) coef(lm(Postwt ~ Prewt + Treat + offset(Prewt), data = d))
cs3 <- cheap_bootstrap(fun=x, b = 20, data = anorexia, parallelize = TRUE, cores = 2)
summary(cs3)

## example with custom function returning coefficients 
## example from ATE-package
library(survival)
library(riskRegression)
set.seed(10)
#### Survival settings  ####
#### ATE with Cox model ####

## generate data

n <- 100
dtS <- sampleData(n, outcome="survival")
dtS$time <- round(dtS$time,1)
dtS$X1 <- factor(rbinom(n, prob = c(0.3,0.4) , size = 2), labels = paste0("T",0:2))

## estimate the Cox model
fit <- coxph(formula = Surv(time,event)~ X1+X2,data=dtS,y=TRUE,x=TRUE)
## fitter function which returns a named vector 
ate_fit_fun <- function(d) {
  ## main fitter function; ignore output
  invisible(capture.output(ate_fit <- summary(ate(fit, 
                                                  data = d, 
                                                  treatment = "X1", 
                                                  times = 5:8, 
                                                  se = FALSE))))
  ## extract the point estimates for risk difference
  res <- ate_fit$diffRisk$estimate
  ## name the point estimates
  names(res) <- paste0("ATE ",
                       ate_fit$diffRisk$A,
                       "-",
                       ate_fit$diffRisk$B, 
                       " (t = ",
                       ate_fit$diffRisk$time,")") 
  res
}
set.seed(102)
cs4 <- cheap_bootstrap(ate_fit_fun, b = 5, data = dtS)
summary(cs4)

## example from riskRegression with no existing coef function
set.seed(18)
learndat <- sampleData(200,outcome="binary")
testdat <- sampleData(1200,outcome="binary")

## score logistic regression models
lr1 = glm(Y~X1+X2+X7+X9,data=learndat,family=binomial)
lr2 = glm(Y~X3+X5,data=learndat,family=binomial)
z<-Score(list("LR(X1+X2+X7+X9)"=lr1,
         "LR(X3+X5)"=lr2),
         formula=Y~1,
         data=testdat,
         metrics = "AUC")
coef.Score <- function(x) {
  res <- x$AUC$score$AUC
  names(res) <- paste0("AUC (model = ", x$AUC$score$model, ")")
  res
}
set.seed(102)
cs5 <- cheap_bootstrap(z, b = 20)
summary(cs5)
}